Web refactor: Layer separation

+++ backend/models/produto.py
@@ -1,4 +1,5 @@
-from sqlalchemy import Column, String, Numeric
+from sqlalchemy import Column, String, Float
+
 from backend.models.base_model import BaseModel
 
 
@@ -7,7 +8,7 @@ class Product(BaseModel):
 
     name = Column(String(length=200), nullable=False)
     description = Column(String(length=200), nullable=False)
-    price = Column(Numeric, nullable=False)
+    price = Column(Float, nullable=False)
 
     def __init__(self, name: str, description: str, price: float) -> None:
         self.name = name
@@ -16,4 +17,3 @@ class Product(BaseModel):
 
     def __str__(self) -> str:
         return f"Name: {self.name}, Description: {self.description}, Price: {self.price}"
-
